backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.7.0
pillow==11.3.0
sqlparse==0.5.3
djangorestframework-simplejwt==5.3.1
asgiref==3.9.1


# to download 
# pip install -r requirements.txt


# To create superUser

# first do migrate for acess
# python manage.py migrate
# python manage.py createsuperuser